Updates gem summary / description
We aren't really extending the object, more just adding some methods that we'll
find useful. We could look into using delegation and literally extended the
class, but for now we'll keep it simple. As a side note, ExtendedCSV may not be
the best name, because it implies that we still have all the underlying methods
that the CSV class has, which is not the class. Let's consider alternative
naming options.